reorganization for styled components, for breakdown and reviews separation, AND fixing the import routes for affected files
