+++ client/src/components/reviews/ReviewsMain.jsx
@@ -1,18 +1,16 @@
 import React, {useState} from 'react';
-import Breakdown from './Breakdown.jsx';
-import ReviewsList from './ReviewsList.jsx';
-import styled from 'styled-components';
+import Breakdown from './breakdownComponents/Breakdown.jsx';
+import ReviewsList from './reviewListComponents/ReviewsList.jsx';
+import {Container} from './sharedStyles/sharedStyledComponents.js'
 import data from './apiExample.js'
 
-const Container = styled.div`
-  text-align: center;
-`
+
 
 let Reviews = () => {
   // let [review, setReview] = useState()
   return (<Container>
     <h1>Ratings & Reviews</h1>
-    <Breakdown meta={data.reviewsMeta}/> <ReviewsList reviews={data.reviews} />
+    <Breakdown meta={data.reviewsMeta}/> <ReviewsList reviews={data.reviews}/>
     </Container>);
 }
